vcmi: move ServerSpellCastEnvironment in separate file
In theory, it should not know about CGameHandler
